Gerenciamento de Tarefas 2.0
Transformando o projeto em um arquivo exe
